+++ doc/source/code/tut_artif.py
@@ -29,7 +29,7 @@ netParams.synMechParams['AMPA'] = {'mod': 'Exp2Syn', 'tau1': 0.1, 'tau2': 1.0, '
 
 # Stimulation parameters
 netParams.stimSourceParams['background'] = {'type': 'NetStim', 'interval': 100, 'number': 1e5, 'start': 500, 'noise': 0.5}  # stim using NetStims after 500ms
-netParams.stimTargetParams['bkg->PYR1'] = {'source': 'background', 'conds': {'popLabel': 'PYR1'}, 'sec':'soma', 'loc': 0.5, 'weight': 0.5, 'delay': 1}
+netParams.stimTargetParams['bkg->PYR1'] = {'source': 'background', 'conds': {'pop': 'PYR1'}, 'sec':'soma', 'loc': 0.5, 'weight': 0.5, 'delay': 1}
 
 
 # Cell parameters
@@ -43,20 +43,20 @@ netParams.cellParams['PYR'] = cellParams
 
 # Connections
 netParams.connParams['artif1->PYR1'] = {
-    'preConds': {'popLabel': 'artif1'}, 'postConds': {'popLabel': 'PYR1'},
+    'preConds': {'pop': 'artif1'}, 'postConds': {'pop': 'PYR1'},
     'convergence': 8,
     'weight': 0.005,                    
     'synMech': 'AMPA',                
     'delay': 'uniform(1,5)'}          
 
 netParams.connParams['PYR1->artif2'] = {
-    'preConds': {'popLabel': 'PYR1'}, 'postConds': {'popLabel': 'artif2'},
+    'preConds': {'pop': 'PYR1'}, 'postConds': {'pop': 'artif2'},
     'probability': 0.2,
     'weight': 0.2,                     
     'delay': 'uniform(1,5)'}     
 
 netParams.addConnParams('artif2->artif3',
-    {'preConds': {'popLabel': 'artif2'}, 'postConds': {'popLabel': 'artif3'}, 
+    {'preConds': {'pop': 'artif2'}, 'postConds': {'pop': 'artif3'}, 
     'divergence': 20,
     'weight': 0.05,              
 	'delay': 3})